// share/constants/enums.ts
export enum EnumFields {
    Field__InternalName_Priorities = "TestingPriority",
}
export enum EnumsList {
    List_InternalName_Tickets = "Tickets",
    List_Ticketing= "Tickets",
    List_InternalName_Departments = "Departments",
    List_Departments= "Departments",
}

export enum EnumsModels {
    Ticket = "ticket",
    Employee = "employee",
    infoTask = "infoTask",
    responseInfoTask = "responseInfoTask",
    department = "department",
    deadline = "deadline",
    status = "status",
    priority = "priority",
}

export enum EnumsStatusOptions {
    Started = "Started",
    InProgress = "In Progress",
    Finished = "Finished",
}

export enum EnumsTypeTags {
    TAG_FIELD_LOOKUP = "TAG_FIELD_LOOKUP",
    TAG_FIELD_TEXT = "TAG_FIELD_TEXT",
    TAG_FIELD_CHOICE = "TAG_FIELD_CHOICE",
    TAG_FIELD_DATE = "TAG_FIELD_DATE",
}

export enum EnumsI18N {
    TEXTFIELD_TITLE = 'TEXTFIELD_ITEM',
    TEXTFIELD_EMPLOYEE = 'TEXTFIELD_EMPLOYEE',
    TEXTFIELD_MESSAGE_TICKET = 'TEXTFIELD_MESSAGE_TICKET',
    TEXTFIELD_MESSAGE_TICKET_RESPONSE = 'TEXTFIELD_MESSAGE_TICKET_RESPONSE',
    TEXTFIELD_DEPARTMENT = 'TEXTFIELD_DEPARTMENT',
    TEXTFIELD_DEADLINE = 'TEXTFIELD_DEADLINE',
    TEXTFIELD_STATUS = 'TEXTFIELD_STATUS',
    TEXTFIELD_PRIORITY = 'TEXTFIELD_PRIORITY',
    BTN_SAVE_TEXT = 'BTN_SAVE_TEXT',
    BTN_EXIT_TEXT = 'BTN_EXIT_TEXT',
    PLACEHOLDER_TEXT_DEPARTMENT = 'PLACEHOLDER_TEXT_DEPARTMENT',
    PLACEHOLDER_TEXT_STATUS = 'PLACEHOLDER_TEXT_STATUS',
    PLACEHOLDER_TEXT_PRIORITY = 'PLACEHOLDER_TEXT_PRIORITY',
    MESSAGE_ITEM_SAVED_OK = 'MESSAGE_ITEM_SAVED_OK',
    MESSAGE_ITEM_SAVED_ERROR = 'MESSAGE_ITEM_SAVED_ERROR',
}